backend/requirements.txt
flask>=2.3
flask-cors>=3.0
python-dotenv>=1.0
pymysql>=1.1
markdown>=3.4
openai>=1.0
langchain>=0.1.0
langchain-openai>=0.1.0
